+++ assets/outputs/project_v0/version_02/image_01.png
